Migrate Hero components to their own directory
