+++ src/constants.js
@@ -1,7 +1,8 @@
 const imagePath = `${window.location.origin}${window.location.pathname}static/media`;
 
 const gtagCategory = 'Component Interaction';
+const gtagTabCategory = 'Demo Tab Interaction';
 const gtagButtonAction = 'button_click';
 const gtagTextFieldAction = 'text_field_click';
 
-export {imagePath, gtagButtonAction, gtagCategory, gtagTextFieldAction};
+export {imagePath, gtagButtonAction, gtagCategory, gtagTabCategory, gtagTextFieldAction};
